# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockIQ — WMS API",
    "version": "1.0.0",
    "description": "API REST do StockIQ (WMS / ERP Lite).\n\n**Autenticação:** faça `POST /api/auth/login`, copie o `token` retornado e clique em **Authorize** (esquema `bearerAuth`).\n\n**Perfis:** `admin` > `supervisor` > `operator` > `viewer`. Cada operação lista os perfis permitidos.\n\n**Erros:** respostas de erro seguem `{ \"message\": string }`; validações retornam 422 com `errors[]`."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LotExpirationReport": {
        "type": "object",
        "properties": {
          "windowDays": {
            "type": "integer",
            "example": 90
          },
          "summary": {
            "type": "object",
            "properties": {
              "expired": {
                "type": "integer",
                "example": 2
              },
              "expiringSoon": {
                "type": "integer",
                "example": 3
              },
              "valid": {
                "type": "integer",
                "example": 4
              },
              "quarantine": {
                "type": "integer",
                "example": 1
              }
            }
          },
          "expired": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lot"
            }
          },
          "expiringSoon": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lot"
            }
          },
          "quarantine": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lot"
            }
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "company": {
            "$ref": "#/components/schemas/Company"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          }
        }
      },
      "UpdateSettingsRequest": {
        "type": "object",
        "properties": {
          "company": {
            "$ref": "#/components/schemas/Company",
            "description": "Partial update"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences",
            "description": "Partial update"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Token ausente, inválido ou expirado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Perfil sem permissão para esta operação",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Erro de validação do corpo/parâmetros",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/settings": {
      "get": {
        "summary": "Get company data and system preferences",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "Current settings (defaults are returned for values never saved)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update company data and/or preferences",
        "description": "Partial update, `company` and `preferences` are merged with the stored values.\n\n**Perfis permitidos:** admin.",
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/reports/movements": {
      "get": {
        "summary": "Movement report grouped by type and product",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movement totals by type and top products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementReport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/api/reports/lots": {
      "get": {
        "summary": "Lot expiration report",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 90
            },
            "description": "Window in days to flag lots expiring soon"
          }
        ],
        "responses": {
          "200": {
            "description": "Expired, expiring soon, valid, and quarantine lots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotExpirationReport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/reports/warehouse": {
      "get": {
        "summary": "Warehouse occupancy report",
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "Addresses grouped by status with capacity and occupancy rate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseReport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/reports/inventory": {
      "get": {
        "summary": "Inventory count report with divergences and accuracy",
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "All inventory counts with summary statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryReport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/reports/suppliers": {
      "get": {
        "summary": "Supplier report — entries and purchase value per supplier",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Purchase totals grouped by supplier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierReport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/customers": {
      "get": {
        "summary": "List customers with search and pagination",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name, tradeName, cnpj or email"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of customers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCustomer"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "tradeName",
                  "cnpj",
                  "email",
                  "city",
                  "state"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "tradeName": {
                    "type": "string"
                  },
                  "cnpj": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "contactName": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "**Perfis permitidos:** admin, supervisor."
      }
    },
...
    "/api/audit": {
      "get": {
        "summary": "List audit logs",
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "in": "query",
            "name": "entity",
            "schema": {
              "type": "string"
            },
            "description": "Filter by entity type (e.g. Product, User)"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by user, action, entity or record name"
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAuditLog"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "**Perfis permitidos:** admin, supervisor."
      }
    },
...
    "/api/analytics": {
      "get": {
        "summary": "Analytics overview (forecast, ABC×XYZ matrix, purchase suggestions, insights)",
        "description": "Computed on demand from the movement history. Forecast uses a least-squares linear trend over monthly exits; XYZ uses the coefficient of variation of monthly demand.\n",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "historyMonths",
            "schema": {
              "type": "integer",
              "minimum": 3,
              "maximum": 24,
              "default": 6
            }
          },
          {
            "in": "query",
            "name": "forecastMonths",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 6,
              "default": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsOverview"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/alerts": {
      "get": {
        "summary": "Get all active alerts (stock + expiry) with per-user read status",
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 30
            },
            "description": "How many days ahead to look for expiring lots"
          }
        ],
        "responses": {
          "200": {
            "description": "Combined alert list with unread count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/api/lots/alerts": {
      "get": {
        "summary": "Get lots expiring soon (default 30 days)",
        "tags": [
          "Lots"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lots expiring soon",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lot"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/api/dashboard": {
      "get": {
        "summary": "Get dashboard summary (KPIs)",
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "Summary with products, movements, users, warehouse stats and stock value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}